requirements.txt
addict==2.4.0
anyio==4.4.0
astor==0.8.1
autodistill==0.1.28
autodistill-grounding-dino==0.1.4
autodistill-yolov8==0.1.4
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.2
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
fonttools==4.53.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle-gpu==2.6.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycocotools==2.0.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.5
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rf-groundingdino==0.1.2
rf-segment-anything==1.0
roboflow==1.1.37
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
sentry-sdk==2.12.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
supervision==0.22.0
sympy==1.13.1
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.7.24
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yapf==0.40.2
zipp==3.19.2
PyMuPDF==1.24.9
